pyproject.toml
[project]
name = "subscription-service"
version = "0.1.0"
requires-python = "==3.13.11"
dependencies = [
    "fastapi[standard]==0.128.0",
    "orjson==3.11.5",
    "gunicorn==23.0.0",
    "pydantic-settings==2.12.0",
    "psycopg[binary]==3.3.2",
    "typer==0.21.1",
    "asgi-lifespan==2.1.0",
    "aiokafka==0.13.0",
    "sqladmin>=0.16.0",
    "aiohttp==3.9.3",
    "python-jose[cryptography]==3.3.0",
    "sqlalchemy[asyncio]==2.0.36",
    "alembic==1.13.2",
    "asyncpg>=0.29",
]

[dependency-groups]
dev = [
    "mypy==1.19.1",
    "ruff==0.14.9",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "asgi-lifespan==2.1.0",
    "pytest-httpx==0.36.0",
]
